Add .circleci/config.yml
It can be difficult to trust that the artifact stored on a Git release is the
artifact built by that commit. EVE being as hard-core as it is, the risk of a
trojan is above nil.  By building the application in a visible CI tool we can
be more confident the result is based on the code.

Further, this allows someone to fork the repo and create a build of their own
(using CircleCI) without locally installing development tools.